% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_odd.R
\name{is_odd}
\alias{is_odd}
\title{Check if a number is odd}
\usage{
is_odd(x)
}
\arguments{
\item{x}{A numeric vector or integer to check.}
}
\value{
A logical vector: \code{TRUE} if the number is odd, \code{FALSE} otherwise.
}
\description{
This function accepts a number (or a vector of numbers) and determines
if it is odd. It returns TRUE for odd numbers and FALSE for even numbers.
}
\examples{
is_odd(3)
is_odd(4)
is_odd(c(1, 2, 3, 5, 8))

}
